*PROTOCOL CHANGE*
the Player message protocol now includes a SYNCH message that is attached to
the end of all of the data messages in each round of data.  the SYNCH message
has its own new type (5) and is addressed from the Player device (1:0).
new datamodes are forthcoming.

C,C++, and Tcl clients have been updated.  haven't bothered with LISP client
yet.


git-svn-id: svn://svn.code.sf.net/p/playerstage/svn/code/player/trunk@615 9fbd2ccc-adb5-4725-87dd-eba157860cdb